Add permissions in share form and conditional rendering to note menu

+++ backend/database_queries.py
@@ -62,34 +62,29 @@ def does_user_exist(cursor: Cursor, user_id: str) -> bool:
 
 #zmienione
 
-# Sprawdzenie, czy użytkownik już istnieje w bazie danych
 def check_existing_user(cursor, user_id: str):
     cursor.execute("SELECT * FROM Users WHERE id = %s", (user_id,))
     return cursor.fetchone()
 
-# Zapisanie nowego użytkownika w bazie danych
 def create_keycloak_user(cursor, user: KeycloakUserCreate):
     cursor.execute("""
         INSERT INTO Users (id, name, last_name, email)
         VALUES (%s, %s, %s, %s);
     """, (user.userId, user.firstName, user.lastName, user.email))
     
-    cursor.connection.commit()  # Zatwierdzamy zmiany w bazie danych
+    cursor.connection.commit()
 
-    # Pobieramy ID zapisanego użytkownika (po wstawieniu)
     cursor.execute("SELECT id FROM Users WHERE id = %s", (user.userId,))
     return cursor.fetchone()[0]  # Zwracamy ID użytkownika
 
 
 def save_user_to_db(cursor: Cursor, user_id: str, name: str, lastname: str, email: str) -> bool:
-    # Sprawdzenie, czy użytkownik już istnieje
     cursor.execute("SELECT id FROM Users WHERE id = %s", (user_id,))
     rows = cursor.fetchall()
 
     if len(rows) > 0:
-        return False  # Użytkownik już istnieje, nie zapisujemy
+        return False
 
-    # Zapisujemy nowego użytkownika
     query = """
     INSERT INTO Users (id, name, last_name, email)
     VALUES (%s, %s, %s, %s);
@@ -102,7 +97,6 @@ def save_user_to_db(cursor: Cursor, user_id: str, name: str, lastname: str, emai
 
 def decode_token_and_save_user(cursor: Cursor, token: str) -> bool:
     try:
-        # Dekodowanie tokenu (zakładając, że nie weryfikujemy podpisu w tej funkcji)
         decoded = jwt.decode(token, options={"verify_signature": False})
         
         user_id = decoded.get("sub")
@@ -110,12 +104,10 @@ def decode_token_and_save_user(cursor: Cursor, token: str) -> bool:
         lastname = decoded.get("family_name")
         email = decoded.get("email")
 
-        # Sprawdzamy, czy wszystkie wymagane dane istnieją
         if not user_id or not name or not lastname or not email:
             print("Brak wymaganych danych w tokenie.")
             return False
 
-        # Zapisujemy użytkownika w bazie danych
         if save_user_to_db(cursor, user_id, name, lastname, email):
             print(f"Użytkownik {name} {lastname} zapisany do bazy.")
             return True
@@ -164,47 +156,45 @@ def get_note(cursor: Cursor, note_id: int) -> str:
     return cursor.fetchone()[0]
 
 
-def delete_note(cursor: Cursor, note_id: int):
-    cursor.execute(
-        "DELETE FROM UserNotes WHERE note_id = %s;", 
-        (note_id))
-    cursor.execute(
-        "DELETE FROM CategoryNotes WHERE note_id = %s;", 
-        (note_id))
-    cursor.execute(
-        "DELETE FROM Notes WHERE id = %s;", 
-        (note_id))
+def delete_note(cursor: Cursor, note_id: int, user_id: str):
+    query_check_owner = """
+    SELECT permission
+    FROM UserNotes
+    WHERE note_id = %s AND user_id = %s;
+    """
+    cursor.execute(query_check_owner, (note_id, user_id))
+    result = cursor.fetchone()
 
-#zmienione
-# def get_user_notes(cursor: Cursor, user_id: str) -> list[tuple[int, str]]:
-#     query1 = """
-# SELECT Notes.id, Notes.title FROM Notes
-# INNER JOIN UserNotes ON Notes.id = UserNotes.note_id
-# INNER JOIN Users ON UserNotes.user_id = %s
-# WHERE UserNotes.permission > 0;
-# """
-    
-#     cursor.execute(query1, (user_id))
-#     notes1 = cursor.fetchall()
+    if result is None:
+        return {"message": "Brak uprawnień do usunięcia notatki."}
 
-#     query2 = """
-# SELECT Notes.id, Notes.title FROM Notes
-# INNER JOIN CategoryNotes ON Notes.id = CategoryNotes.note_id
-# INNER JOIN Categories ON CategoryNotes.category_id = Categories.id
-# INNER JOIN UserCategories ON Categories.id = UserCategories.category_id
-# INNER JOIN Users ON UserCategories.user_id = %s
-# WHERE CategoryNotes.permission > 0;
-# """
+    permission = result[0]
 
-#     cursor.execute(query2, (user_id))
-#     notes2 = cursor.fetchall()
-    
-#     return list(dict.fromkeys(notes1 + notes2))
+    if permission == 3:
+        query_delete_access = """
+        DELETE FROM UserNotes WHERE note_id = %s;
+        """
+        cursor.execute(query_delete_access, (note_id,))
 
-#to działa, ale dla userów z id liczbowym
-def get_user_notes(cursor: Cursor, user_id: str) -> list[tuple[int, str, str]]:
+        query_delete_note = """
+        DELETE FROM Notes WHERE id = %s;
+        """
+        cursor.execute(query_delete_note, (note_id,))
+
+        return {"message": "Notatka została całkowicie usunięta."}
+
+    query_delete_access = """
+    DELETE FROM UserNotes WHERE note_id = %s AND user_id = %s;
+    """
+    cursor.execute(query_delete_access, (note_id, user_id))
+
+    return {"message": "Usunięto dostęp do notatki."}
+
+
+
+def get_user_notes(cursor: Cursor, user_id: str) -> list[tuple[int, str, int, str]]:
     query1 = """
-SELECT Notes.id, Notes.title, Categories.name AS category
+SELECT Notes.id, Notes.title, UserNotes.permission, Categories.name AS category
 FROM Notes
 INNER JOIN UserNotes ON Notes.id = UserNotes.note_id
 LEFT JOIN CategoryNotes ON Notes.id = CategoryNotes.note_id
@@ -216,7 +206,7 @@ WHERE UserNotes.permission > 0;
     notes1 = cursor.fetchall()
 
     query2 = """
-SELECT Notes.id, Notes.title, Categories.name AS category
+SELECT Notes.id, Notes.title, CategoryNotes.permission, Categories.name AS category
 FROM Notes
 INNER JOIN CategoryNotes ON Notes.id = CategoryNotes.note_id
 LEFT JOIN Categories ON CategoryNotes.category_id = Categories.id
@@ -224,19 +214,19 @@ INNER JOIN UserCategories ON Categories.id = UserCategories.category_id
 INNER JOIN Users ON UserCategories.user_id = %s
 WHERE CategoryNotes.permission > 0;
 """
-
     cursor.execute(query2, (user_id))
     notes2 = cursor.fetchall()
 
     # Łączymy oba zestawy wyników i usuwamy duplikaty
     all_notes = list(dict.fromkeys(notes1 + notes2))
 
-    # Dodajemy kategorię lub "Prywatny" jeśli brak kategorii
+    # Zwracamy id, title, permission oraz category
     return [
         {
             "id": note[0],
             "title": note[1],
-            "category": note[2] if note[2] else "Prywatny"
+            "permission": note[2],  # Nowa kolumna - permission
+            "category": note[3] if note[3] else "Prywatny"
         }
         for note in all_notes
     ]
@@ -255,33 +245,13 @@ def create_category(cursor: Cursor, category: Category):
 INSERT INTO Categories (name)
 VALUES (%s);
 """
-    # Zapisywanie kategorii
     cursor.execute(query, (category.name))
 
-#zmienione
-# def get_categories(cursor: Cursor) -> list[tuple[int, str]]:
-#     cursor.execute("SELECT * FROM Categories")
-#     return list(cursor.fetchall())
 
 #stare
 def get_categories(cursor: Cursor) -> list[tuple[int, str]]:
     cursor.execute("SELECT id, name FROM Categories")
     return list(cursor.fetchall())
-
-#nowe
-# def get_categories(cursor: Cursor) -> list[tuple[int, str]]:
-#     cursor.execute("""
-# SELECT
-#     n.id AS note_id,
-#     n.title,
-#     n.content,
-# GROUP_CONCAT(c.name) AS categories
-# FROM Notes n
-# LEFT JOIN CategoryNotes cn ON n.id = cn.note_id
-# LEFT JOIN Categories c ON cn.category_id = c.id
-# WHERE n.id IN (SELECT note_id FROM UserNotes WHERE user_id = ?)
-# GROUP BY n.id;""")
-#     return list(cursor.fetchall())
 
 
 #endregion
